refractor: separated socket.io
